requirements.txt
django
django-rest-frameworks
drf-spectacular